+++ Source/Android/app/src/main/res/layout/fragment_ingame_menu.xml
@@ -101,6 +101,11 @@
                 android:text="@string/emulation_change_disc"
                 style="@style/InGameMenuOption" />
 
+            <Button
+                android:id="@+id/menu_skylanders"
+                android:text="@string/emulate_skylander_portal"
+                style="@style/InGameMenuOption" />
+
         </LinearLayout>
 
     </ScrollView>
