+++ Source/Android/app/src/main/res/layout/list_item_skylander_slot.xml
@@ -1,0 +1,61 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout android:id="@+id/root"
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:padding="@dimen/spacing_medlarge">
+
+    <Button
+        android:id="@+id/button_create_skylander"
+        style="?attr/materialIconButtonFilledTonalStyle"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_centerVertical="true"
+        android:layout_marginEnd="@dimen/spacing_small"
+        android:layout_toStartOf="@id/button_load_skylander"
+        android:contentDescription="@string/create_skylander"
+        android:tooltipText="@string/create_skylander"
+        app:icon="@drawable/ic_add" />
+
+    <Button
+        android:id="@+id/button_load_skylander"
+        style="?attr/materialIconButtonFilledTonalStyle"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_centerVertical="true"
+        android:layout_marginEnd="@dimen/spacing_small"
+        android:layout_toStartOf="@+id/button_clear_skylander"
+        android:contentDescription="@string/load_skylander"
+        android:tooltipText="@string/load_skylander"
+        app:icon="@drawable/ic_load" />
+
+    <Button
+        android:id="@+id/button_clear_skylander"
+        style="?attr/materialIconButtonFilledTonalStyle"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_alignParentEnd="true"
+        android:layout_centerVertical="true"
+        android:layout_marginEnd="@dimen/spacing_large"
+        android:contentDescription="@string/remove_skylander"
+        android:tooltipText="@string/remove_skylander"
+        app:icon="@drawable/ic_clear" />
+
+    <com.google.android.material.textview.MaterialTextView
+        android:id="@+id/text_skylander_name"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_alignParentStart="true"
+        android:layout_centerVertical="true"
+        android:layout_marginBottom="@dimen/spacing_medlarge"
+        android:layout_marginEnd="@dimen/spacing_large"
+        android:layout_marginStart="@dimen/spacing_large"
+        android:layout_toStartOf="@+id/button_create_skylander"
+        android:textAlignment="viewStart"
+        android:textColor="?attr/colorOnSurface"
+        android:textSize="16sp"
+        tools:text="Wii Remote with Motion Plus Pointing" />
+
+</RelativeLayout>
