+++ Source/Core/Core/IOS/USB/Emulated/Skylander.h
@@ -19,6 +19,7 @@ constexpr u8 MAX_SKYLANDERS = 16;
 
 namespace IOS::HLE::USB
 {
+extern const std::map<const std::pair<const u16, const u16>, const char*> list_skylanders;
 class SkylanderUSB final : public Device
 {
 public:
@@ -95,6 +96,7 @@ public:
   bool CreateSkylander(const std::string& file_path, u16 sky_id, u16 sky_var);
   bool RemoveSkylander(u8 sky_num);
   u8 LoadSkylander(u8* buf, File::IOFile in_file);
+  std::pair<u16, u16> CalculateIDs(const std::array<u8, 0x40 * 0x10>& file_data);
 
 protected:
   std::mutex sky_mutex;
